+++ app/charting/fetchCharts.go
@@ -1,4 +1,4 @@
-package charts
+package charting
 
 import "fmt"
